Add files via upload
2 Simple Questionnaire application.
the first creates an HTML page with a form that displays a drop list with four options. When the form is submitted, a PHP file uses if...elseif...else blocks to generate a response to the option selected.

the second program is the exact same but the responses are generated with a case switch suite.