Add files via upload
This app uses a PDO to retrieve and display records in a database table. The database (pcparts) is provided. The app initially displays a web page with a form for specifying a city. Submitting the form should output all fields in the customers table for customers in the specified city.

once you've uploaded the DB create a pcparts user named parts_dude with password P@rtzer. Give this user privileges at minimum to  select.